# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SubGlance API
  version: v1
  description: |
    The HTTP API of SubGlance, a self-hosted uptime monitor.

    The web UI is a client of this API and has no private endpoints: anything
    the dashboard can do, a script can do (product principle 4).

    Authentication accepts either a browser session cookie or a bearer API
    token; both resolve to the same user with the same rights. Routes are
    grouped by what they require — `public`, `read` (any authenticated user,
    including viewers), `write` (editor or admin) and `admin`. Every operation
    records its level in the `x-subglance-access` extension, which is checked
    against the server's route table by an automated drift test.

    Timestamps are RFC 3339 / ISO 8601 in UTC. Errors always share one shape:
    a JSON object with a single `error` string.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: /
    description: The SubGlance instance itself; the API is served from its own origin.
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/v1/status-pages:
    get:
      tags: [Status pages]
      operationId: listStatusPages
      summary: List status pages
      description: |
        Administrators only. Every page with its settings and entries, oldest
        first, including disabled pages. This is the operator's view: it
        carries monitor ids and is never what a visitor sees.
      responses:
        '200':
          description: Every status page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusPageList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenAdmin'
        '500':
          description: The pages could not be read.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-subglance-access: admin
    post:
      tags: [Status pages]
      operationId: createStatusPage
      summary: Create a status page
      description: |
        Administrators only. A new page is off (`enabled: false`) and asks
        search engines not to index it unless the request says otherwise, so
        saving a draft publishes nothing. Add monitors to it with
        `PUT /api/v1/status-pages/{slug}/entries`.

        The slug is trimmed and lowercased; `api`, `assets` and `fonts` are
        reserved. An empty time zone becomes `UTC`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusPageRequest'
      responses:
        '201':
          description: The page as stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusPage'
        '400':
          description: |
            Malformed or unknown-field JSON, or a setting that fails
            validation. A validation failure carries `field`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: slug "api" is reserved
                field: slug
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenAdmin'
        '409':
          description: Another page already uses this slug, in any letter case.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: another status page already uses this slug
                field: slug
        '500':
          description: The page could not be stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-subglance-access: admin
  /api/v1/status-pages/{slug}:
    parameters:
      - $ref: '#/components/parameters/StatusPageSlug'
    get:
      tags: [Status pages]
      operationId: getPublicStatusPage
      summary: Read a published status page (public)
      description: |
        What a visitor sees, as JSON: the same answer the HTML page at
        `GET /status/{slug}` is drawn from. No credentials are needed or
        read. Monitors appear under their public names and random keys only;
        internal names, ids, targets, tags, check results and incident causes
        never do.

        A slug that names no page and a page that is switched off get the
        same `404`, byte for byte, so the answer does not reveal which slugs
        exist.

        Rate limited per client address (5 per second, bursts of 20) and in
        total (50 per second), before the page is looked up; behind a proxy
        the address comes from `--trusted-proxies`. The answer is rendered at
        most once per 30 seconds and carries `Cache-Control: public` with the
        seconds left, and an `ETag` for `If-None-Match`. A request served
        from that cache is not counted against the limits. Switching a page
        off therefore shows within 30 seconds, not instantly.

        `X-Robots-Tag: noindex` is sent unless the page allows indexing.
      security: []
      parameters:
        - name: If-None-Match
          in: header
          required: false
          description: An `ETag` from an earlier answer. A match returns `304`.
          schema:
            type: string
      responses:
        '200':
          description: The page.
          headers:
            ETag:
              $ref: '#/components/headers/PublicStatusPageETag'
            Cache-Control:
              $ref: '#/components/headers/PublicStatusPageCacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicStatusPage'
        '304':
          description: The caller's copy, named by `If-None-Match`, is current.
        '404':
          $ref: '#/components/responses/StatusPageNotFound'
        '429':
          $ref: '#/components/responses/PublicStatusPageLimited'
        '503':
          description: |
            The page could not be built. The body is fixed; the cause is in
            the server log.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-subglance-access: public
    put:
      tags: [Status pages]
      operationId: updateStatusPage
      summary: Replace a status page's settings
      description: |
        Administrators only. Every setting is replaced: an omitted boolean is
        `false`, so a client that forgets `enabled` takes the page offline
        rather than publishing it. The slug may change; the page is then
        found under the new one. Entries are left alone.

        Switching a page off takes effect within the public page's cache
        interval of 30 seconds, not instantly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusPageRequest'
      responses:
        '200':
          description: The page after the update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusPage'
        '400':
          description: |
            Malformed or unknown-field JSON, or a setting that fails
            validation. A validation failure carries `field`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenAdmin'
        '404':
          $ref: '#/components/responses/StatusPageNotFound'
        '409':
          description: Another page already uses the new slug.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: The page could not be updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-subglance-access: admin
    delete:
      tags: [Status pages]
      operationId: deleteStatusPage
      summary: Delete a status page
      description: |
        Administrators only. The page and its entries go; the monitors stay.
        Public keys of the removed entries are never issued again.
      responses:
        '204':
          description: The page was deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenAdmin'
        '404':
          $ref: '#/components/responses/StatusPageNotFound'
        '500':
          description: The page could not be deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-subglance-access: admin
  /api/v1/status-pages/{slug}/entries:
    parameters:
      - $ref: '#/components/parameters/StatusPageSlug'
    put:
      tags: [Status pages]
      operationId: setStatusPageEntries
      summary: Replace the monitors on a status page
      description: |
        Administrators only. The list becomes the page's complete set of
        entries, in the order given; `[]` removes every monitor. The change is
        all or nothing.

        Every entry needs a `display_name`: it is the only name a visitor
        sees, and there is no fallback to the monitor's own name. A monitor
        that stays on the page keeps its `public_key` through renames and
        reorders; one that is added gets a new key.

        On a page that selects by tag, only entries whose monitor carries the
        tag are shown; `unnamed_monitor_ids` in the response lists the tagged
        monitors that still need an entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusPageEntriesRequest'
      responses:
        '200':
          description: The page with its new entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusPage'
        '400':
          description: |
            Malformed or unknown-field JSON, a missing `entries` list, a
            missing or too long display name, a monitor listed twice, more
            than 200 entries, or an unknown monitor id. Carries
            `field: entries` for all but malformed or unknown-field JSON.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: display name of monitor 3 must be 1-80 characters
                field: entries
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenAdmin'
        '404':
          $ref: '#/components/responses/StatusPageNotFound'
        '500':
          description: The entries could not be stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-subglance-access: admin
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-subglance_session
      description: |
        The browser session cookie, set by `POST /api/v1/auth/login`, by
        `POST /api/v1/setup` and by `POST /api/v1/auth/password`. It is
        HttpOnly, SameSite=Lax, Path=/ and valid for seven days of inactivity
        (refreshed on use).

        The `__Host-` prefix is a browser-enforced guarantee that the cookie is
        Secure, has Path=/ and carries no Domain attribute — that last part is
        what stops a subdomain from writing a cookie this origin would accept.
        Over plain HTTP browsers refuse Secure cookies, so an instance without
        TLS falls back to the name `subglance_session`; a self-hoster on a LAN
        must still be able to log in. Both names are accepted on requests.

        Cookie-authenticated state-changing requests additionally pass a CSRF
        check based on `Sec-Fetch-Site` (falling back to an `Origin` match for
        non-browser clients). Bearer-token calls skip it, because a browser
        will not attach a bearer token to a cross-site request.
    bearerToken:
      type: http
      scheme: bearer
      description: |
        An API token in `Authorization: Bearer <token>`, created via
        `POST /api/v1/tokens`. Intended for scripts and CI.

        A token acts with the lower of its own role and the current role of
        the user who created it. When an `Authorization`
        header is present it is the only credential considered: a malformed one
        is rejected rather than falling back to the session cookie.
  parameters:
...
    StatusPageSlug:
      name: slug
      in: path
      required: true
      description: The page's slug, matched without regard to letter case.
      schema:
        type: string
        pattern: '^[A-Za-z0-9][A-Za-z0-9-]{0,62}$'
      example: status
...
  headers:
...
    PublicStatusPageETag:
      description: |
        A strong entity tag for this exact answer. Send it back as
        `If-None-Match` to be told `304` while it is current.
      schema:
        type: string
      example: '"3q2-7wEAAAAAAAAAAAAAAAAA"'
    PublicStatusPageCacheControl:
      description: '`public, max-age=N`, N being the seconds left of the 30-second render cache.'
      schema:
        type: string
      example: public, max-age=30
...
  responses:
    Unauthorized:
      description: No valid credentials were presented.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: authentication required
...
    ForbiddenAdmin:
      description: The caller is authenticated but is not an administrator.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: this action requires an administrator
...
    StatusPageNotFound:
      description: No status page with that slug.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: status page not found
    PublicStatusPageLimited:
      description: |
        Too many requests from this address, or in total. `Retry-After`
        says when to try again.
      headers:
        Retry-After:
          schema:
            type: integer
          example: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: too many requests; try again in a moment
...
  schemas:
...
    Error:
      type: object
      description: |
        The single error shape used by every endpoint. `error` is always
        present and is always a human-readable sentence.

        `field` is present only when the rejection is about one named field of
        the request body, and it carries the JSON name that was sent, so a
        client can place the message next to the input that caused it instead
        of matching on the wording of `error`. Anything not about a single
        field — malformed JSON, a rate limit, a missing session, a server
        failure — omits it.
      required: [error]
      properties:
        error:
          type: string
          description: A human-readable explanation of what went wrong.
          example: monitor not found
        field:
          type: string
          description: |
            The JSON name of the request body field the error is about, when
            the error is about exactly one. Absent otherwise; never empty.
          example: target
      additionalProperties: false
...
    StatusPageRequest:
      type: object
      required: [slug, title, selection]
      additionalProperties: false
      properties:
        slug:
          type: string
          description: |
            Lowercase letters, digits and dashes, starting with a letter or
            digit, at most 63 characters, so it is also a valid DNS label.
            Capitals are lowercased. `api`, `assets`, `fonts` and `logos` are
            reserved.
          pattern: '^[A-Za-z0-9][A-Za-z0-9-]{0,62}$'
          example: status
        title:
          type: string
          description: Shown to visitors. No line breaks or tabs.
          minLength: 1
          maxLength: 120
          example: Acme services
        description:
          type: string
          description: Optional plain text shown under the title.
          maxLength: 500
        timezone:
          type: string
          description: |
            The IANA zone the page shows times in, chosen for its audience.
            Empty means `UTC`; `Local` is refused.
          example: Europe/Amsterdam
        selection:
          type: string
          enum: [monitors, tag]
          description: |
            `monitors` shows exactly the monitors with an entry. `tag` shows
            the entries whose monitor carries `tag_key`=`tag_value`.
        tag_key:
          type: string
          description: 'Required for `selection: tag`, empty otherwise. Normalised like a monitor tag.'
        tag_value:
          type: string
          description: 'Required for `selection: tag`, empty otherwise.'
        indexable:
          type: boolean
          description: Lets search engines index the page. Omitted means false.
        enabled:
          type: boolean
          description: Makes the page reachable. Omitted means false.
        language:
          type: string
          enum: ['', en, nl]
          description: The language of the page's fixed texts. Empty or omitted means `en`.
        accent:
          type: string
          description: |
            The colour of the page title as `#rrggbb`, or empty for the
            default ink. Lowercased on save. It must reach a contrast of 3:1
            against the page in both the dark and the light theme, because
            the page follows each visitor's own theme; a colour that does not
            is refused with both measured ratios (`field: accent`).
          example: '#3b82f6'
        hide_credit:
          type: boolean
          description: |
            Hides "Monitored with SubGlance" in the footer. Omitted means
            false, so the credit stays. The time zone on the same line
            always stays.
...
    StatusPage:
      type: object
      required: [id, slug, title, description, timezone, selection, tag_key, tag_value,
        indexable, enabled, language, accent, hide_credit, logo, created_at, updated_at, entries,
        unnamed_monitor_ids]
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        timezone:
          type: string
        selection:
          type: string
          enum: [monitors, tag]
        tag_key:
          type: string
          description: Empty unless `selection` is `tag`.
        tag_value:
          type: string
          description: Empty unless `selection` is `tag`.
        indexable:
          type: boolean
        enabled:
          type: boolean
        language:
          type: string
          enum: [en, nl]
        accent:
          type: string
          description: '`#rrggbb`, or empty for none.'
        hide_credit:
          type: boolean
        logo:
          description: The uploaded logo, or null.
          oneOf:
            - type: 'null'
            - type: object
              required: [content_type, width, height, bytes, path, updated_at]
              properties:
                content_type:
                  type: string
                  enum: [image/png, image/jpeg, image/webp]
                width:
                  type: integer
                height:
                  type: integer
                bytes:
                  type: integer
                path:
                  type: string
                  description: Where visitors load it while the page is published.
                  example: /status/logos/3f9a0c1d5e7b2a48.png
                updated_at:
                  type: string
                  format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        entries:
          type: array
          description: |
            Every entry in page order, including entries a tag page does not
            show right now because their monitor lost the tag.
          items:
            $ref: '#/components/schemas/StatusPageEntry'
        unnamed_monitor_ids:
          type: array
          description: |
            For a tag page, the monitors that carry the tag but have no entry
            and so are not shown until given a public name. Always empty for
            a page that selects monitors.
          items:
            type: integer
            format: int64
    PublicStatusPage:
      type: object
      description: |
        A status page as its visitors see it (docs/design/status-page.md
        section 1.1). Times are UTC; `timezone` is the zone the HTML page
        shows them in.
      required: [title, description, generated_at, timezone, language, accent, show_credit, logo,
        summary, entries, maintenance, outages]
      properties:
        title:
          type: string
        description:
          type: string
        generated_at:
          type: string
          format: date-time
          description: When this answer was built; it is reused for up to 30 seconds.
        timezone:
          type: string
          example: Europe/Amsterdam
        language:
          type: string
          enum: [en, nl]
          description: The language of the HTML page's fixed texts.
        accent:
          type: string
          description: The title's colour as `#rrggbb`, or empty for the default.
        show_credit:
          type: boolean
          description: Whether the HTML page's footer names SubGlance.
        logo:
          description: |
            The page's logo, or null. The path is on this server and its file
            name is random, never the name it was uploaded under.
          oneOf:
            - type: 'null'
            - type: object
              required: [path, width, height]
              properties:
                path:
                  type: string
                  example: /status/logos/3f9a0c1d5e7b2a48.png
                width:
                  type: integer
                height:
                  type: integer
        summary:
          type: object
          required: [up, degraded, down, unmonitored]
          properties:
            up:
              type: integer
            degraded:
              type: integer
            down:
              type: integer
            unmonitored:
              type: integer
              description: Entries with no data yet or not monitored.
        entries:
          type: array
          items:
            $ref: '#/components/schemas/PublicStatusPageEntry'
        maintenance:
          type: array
          description: Maintenance running now or starting within 7 days.
          items:
            type: object
            required: [starts_at, ends_at, keys]
            properties:
              starts_at:
                type: string
                format: date-time
              ends_at:
                type: string
                format: date-time
              keys:
                type: array
                description: The `key` of every entry the window covers.
                items:
                  type: string
        outages:
          type: array
          description: Confirmed outages in the last 14 days, newest first.
          items:
            type: object
            required: [key, started_at, resolved_at, duration_s]
            properties:
              key:
                type: string
              started_at:
                type: string
                format: date-time
              resolved_at:
                type: [string, 'null']
                format: date-time
              duration_s:
                type: integer
                format: int64
...
    StatusPageList:
      type: object
      required: [pages]
      properties:
        pages:
          type: array
          items:
            $ref: '#/components/schemas/StatusPage'
    StatusPageEntriesRequest:
      type: object
      required: [entries]
      additionalProperties: false
      properties:
        entries:
          type: array
          maxItems: 200
          description: The page's monitors in display order. `[]` removes them all.
          items:
            type: object
            required: [monitor_id, display_name]
            additionalProperties: false
            properties:
              monitor_id:
                type: integer
                format: int64
                minimum: 1
              display_name:
                type: string
                minLength: 1
                maxLength: 80
                description: Trimmed; no line breaks or tabs.
...